api/requirements.txt
clip-anytorch
autokeras==1.0.20
tensorflow==2.9
numpy==1.21
fastapi~=0.73.0
torch==2.0.1
uvicorn[fastapi]
protobuf<=3.20
python-multipart
